pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "songbook-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jeremie Jost", email = "jeremiejost@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.3.0",
    "click>=8.2.1",
    "fsspec>=2025.5.1",
    "functions-framework>=3.8.3",
    "gcsfs>=2025.5.1",
    "google-api-python-client>=2.174.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-pubsub>=2.30.0",
    "google-cloud-trace>=1.14.2",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
    "opentelemetry-propagator-gcp>=1.7.0",
    "opentelemetry-resourcedetector-gcp==1.9.0a0",
    "psutil>=7.0.0",
    "pymupdf>=1.26.1",
    "toml>=0.10.2",
    "pypdf2>=3.0.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.34.1",
    "grpcio>=1.73.1",
    "humanize>=4.12.3",
    "natsort>=8.4.0",
    "unidecode>=1.3.8",
    "numpy>=2.3.1",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "fontra>=0.5.2",
    "loguru>=0.7.0",
    "google-genai>=1.68.0",
    "tenacity>=9.1.4",
]

[project.scripts]
songbook-tools = "generator.cli:cli"

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "psutil>=7.0.0",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "seaborn>=0.13.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["generator*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"generator" = ["fonts/RobotoCondensed-*.ttf"]

[tool.ruff.lint]
extend-select = ["BLE001"] # Catch all exceptions are the plague

[tool.ruff.lint.per-file-ignores]
"generator/cli/__init__.py" = ["E402"] # Early import for warning suppression
